aznfswatchdog should do conntrack reconciliation apart from reconstructing deleted iptable entries
Some kernels have a bug where a conntrack entry gets stuck in SYN_SENT
state. I've seen this resurface sometimes especially when NFS
client reconnects with the same source port (which it always does).
This is a known issue which has been fixed in 5.14 kernels

https://github.com/torvalds/linux/commit/e15d4cdf27cb0c1e977270270b2cea12e0955edd

Here's a discussion thread for kubernetes which details the issue.
https://github.com/kubernetes/kubernetes/issues/101607