Fixing forget_seen assert failure.
We had an assert that fuse should not call forget for an inode for which it
has already called forget. We had a boolean forget_seen in the nfs_inode that
we used to set to true when we used to see a forget call from fuse and we had
an assert that if forget is called forget_seen must be false.
The assert was failing for some cases where we have returned the sme inode as
a result of multiple dir enumeration or some other op, so the lookcupcnt was say 2
and then fuse was calling forget(cnt=1) twice instead of forget(cnt=2).
Fuse is free to do it and we should not assert.

To fix this, now we don't have the forget_seen bool but an integer forget_expected
to count how many forgets to expect. We assert if fuse calls more or less than that.
Also cleaned up the code.