Refactor lookup_dircache(), readdirplus_callback(), readdir_callback() and send_readdir_or_readdirplus_response() to fix an inode leak
Now we hold inode lookupcnt ref for all entries with a valid inode.
This way, send_readdir_or_readdirplus_response() can call decref() for the inodes
it doens't pass to fuse. Previously we were decrementing dircachecnt and there was a possibility
of that becoming 0 while lookupcnt was already 0. Those inodes would be sitting in inode_map
with lookupcnt=dircachecnt=0.