Clear eof, eof_cookie when dir cache is purged
This fixes a bug introduced by a prior change where we removed setting eof
on dircache purge. That was done as we were incorrectly claiming dir shrank
when application/fuse was querying a cookie/offset greater than what prior
readdir call returned.
Now we take care of that condition by only logging the dir shrank warning when
we are querying cookies in order (w/o a gap).