Sparse write must cause inline write
o/w readers after that which read from the sparse section may wrongly return
eof. This happens as those reads are sent to the server as the bytes_chunk
corresponding to the sparse region won't be present and server will return eof
as it doesn't know the actual file size.