Atomically add cookiver and cookie in readdir cache
Currently readdir_callback()/readdirplus_callback() are adding one or more
readdir entries to readdirectory_cache before saving the cookieverf
using a set_cookieverf() call. Since these two are not done atomically, it's
possible that after the cookie/directory_entry is added to the cache, some
other thread may lookup the cache and call send_readdir_or_readdirplus_response().
send_readdir_or_readdirplus_response() asserts that cookieverf stored in the
readdirectory_cache must be non-zero if there's at least one directory_entry.
The reason for the assert is that, since we are returning at least one entry
to fuse, it may ask for more entries and for that we might have to make a
READDIR{PLUS} call to the server, for which we will need a valid cookieverf.

The fix is to update readdirectory_cache::add() to accept cookieverf as an
argument and update cookieverf and entry atomically so that we cannot have
some caller who finds readdirectory_cache with the entry added but cookieverf
not set.

Also added few more useful asserts.