Fix race between flush and unlink.
The race existed as inode->release() used to decrement the opencnt first and then
used to flush any cached data. This gave rise to a possibility where some other thread
calls unlink and it sees the opencnt as 0 and goes ahead with the unlink while the flush
operations are still ongoing. These will reach the server after the file is deleted
and server fails those writes with ENOENT. We currently have an assert that writes
should not fail.

The fix is to decrement the opencnt only after the pending flush operations complete.