Implement 2 phase truncate. (#37)
* Implement 2 phase truncate.

This change gets rid of SCAN_ACTION_TRUNCATE and removes truncate
support from bytes_chunk_cache::scan(). Now scan() cannot remove
membufs which are in use or dirty. truncate() is a separate function
now which specializes in truncating the cache and duly waits for
locked bcs. It assumes that truncate will be called from contexts
which can afford to wait and correctness is more important.

* Enforce mb is locked and inuse inside trim().

* Undo the req to have membuf locked and inuse inside trim()

It's hard to guarantee that, especially from ~bc_iovec().

* Increase max_retry to 200

to properly handle cases where we induce sleep in the write callback

* Do not flush truncated membufs.

---------

Co-authored-by: Nagendra Tomar <natomar@microsoft.com>